test(http): Refactor tests by moving relevant cases from `ApplicationTest` to `ApplicationCoreTest` and removing the obsolete `ApplicationTest` class.
